+++ viewbinding-ktx/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fixed
 
 - Add workaround for the case when fragment view changed without triggering ON_DESTROY
+- Save binding only if fragment view is not destroyed
 
 ## 4.1.2-0 (2021-03-01)